Complete letter generation for each CSV data entry
Added functionality to take the data pulled from the google API and
generate letters for each CSV attendee entry. These letters are then
exported into a directory in the home folder. The letters list each
local representative associated with their given zip code, as well as
their phone numbers if they can be correctly parsed. This is done with
the ERB template created also in the home folder.